requirements.txt
# Core Higgs Audio dependencies
descript-audio-codec
torch
transformers>=4.45.1,<4.47.0
librosa
dacite
boto3==1.35.36
s3fs
torchvision
torchaudio
json_repair
pandas
pydantic
vector_quantize_pytorch
loguru
pydub
ruff==0.12.2
omegaconf
click
langid
jieba
accelerate>=0.26.0

# --- Server dependencies ---
python-dotenv
fastapi
uvicorn[standard]
python-multipart
soundfile
openai>=1.0.0